pyproject.toml
[project]
name = "move3d"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "antlr4-python3-runtime==4.9.3",
    "asttokens==3.0.0",
    "beautifulsoup4==4.13.4",
    "chumpy==0.66",
    "coloredlogs==15.0.1",
    "cycler==0.12.1",
    "decorator==5.2.1",
    "easydict==1.13",
    "executing==2.2.0",
    "filelock==3.18.0",
    "flatbuffers==25.2.10",
    "fsspec==2025.3.2",
    "humanfriendly==10.0",
    "hydra-core==1.3.2",
    "imageio-ffmpeg==0.6.0",
    "ipdb==0.13.13",
    "ipython==9.1.0",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "joblib==1.4.2",
    "kiwisolver==1.4.8",
    "lxml==5.4.0",
    "markupsafe==3.0.2",
    "matplotlib==3.1.1",
    "matplotlib-inline==0.1.7",
    "mpmath==1.3.0",
    "networkx==3.4.2",
    "numpy==2.2.5",
    "nvidia-cublas-cu12==12.6.4.1",
    "nvidia-cuda-cupti-cu12==12.6.80",
    "nvidia-cuda-nvrtc-cu12==12.6.77",
    "nvidia-cuda-runtime-cu12==12.6.77",
    "nvidia-cudnn-cu12==9.5.1.17",
    "nvidia-cufft-cu12==11.3.0.4",
    "nvidia-cufile-cu12==1.11.1.6",
    "nvidia-curand-cu12==10.3.7.77",
    "nvidia-cusolver-cu12==11.7.1.2",
    "nvidia-cusparse-cu12==12.5.4.2",
    "nvidia-cusparselt-cu12==0.6.3",
    "nvidia-nccl-cu12==2.26.2",
    "nvidia-nvjitlink-cu12==12.6.85",
    "nvidia-nvtx-cu12==12.6.77",
    "omegaconf==2.3.0",
    "onnxruntime==1.21.1",
    "onnxruntime-gpu>=1.21.1",
    "opencv-python==4.11.0.86",
    "packaging==25.0",
    "parso==0.8.4",
    "pexpect==4.9.0",
    "pillow==11.2.1",
    "prettytable==3.16.0",
    "prompt-toolkit==3.0.51",
    "protobuf==6.30.2",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pygments==2.19.1",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "pytorch-metric-learning==2.8.1",
    "pyyaml==6.0.2",
    "roma==1.5.2.1",
    "rtmlib==0.0.13",
    "scikit-learn==1.6.1",
    "scipy==1.15.2",
    "setuptools==79.0.1",
    "six==1.17.0",
    "soupsieve==2.7",
    "stack-data==0.6.3",
    "sympy==1.13.3",
    "tensorboardx==2.6.2.2",
    "threadpoolctl==3.6.0",
    "torch==2.7.0",
    "torchaudio==2.7.0",
    "torchvision==0.22.0",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "triton==3.3.0",
    "typing-extensions==4.13.2",
    "wcwidth==0.2.13",
]
